+++ src/test/java/praktikum/IngredientTest.java
@@ -4,25 +4,48 @@ import org.junit.Test;
 import static org.junit.Assert.*;
 import static praktikum.IngredientType.SAUCE;
 
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+import java.util.Arrays;
+import java.util.Collection;
+
+@RunWith(Parameterized.class)
 public class IngredientTest {
+    private IngredientType type;
+    private String name;
+    private float price;
+
+    public IngredientTest(IngredientType type, String name, float price) {
+        this.type = type;
+        this.name = name;
+        this.price = price;
+    }
+
+    @Parameterized.Parameters(name = "Тип ингредиента, соус, цена. Тестовые данные: {0} {1} {2}")
+    public static Collection<Object[]> data() {
+        return Arrays.asList(new Object[][]{
+                {IngredientType.SAUCE, "sour cream", 200},
+                {IngredientType.FILLING, "dinosaur", 200},
+                {IngredientType.FILLING, "sausage", 300}
+        });
+    }
 
     @Test
     public void testGetPrice() {
-        Ingredient ingredient = new Ingredient(SAUCE, "Сырный", 100);
-        float expectedPrice = 100;
-        assertEquals(expectedPrice, ingredient.getPrice(), 0);
+        Ingredient ingredient = new Ingredient(type, name, price);
+        assertEquals(price, ingredient.getPrice(), 0);
     }
 
     @Test
     public void testGetName() {
-        Ingredient ingredient = new Ingredient(SAUCE, "Сырный", 100);
-        String expectedName = "Сырный";
-        assertEquals(expectedName, ingredient.getName());
+        Ingredient ingredient = new Ingredient(type, name, price);
+        assertEquals(name, ingredient.getName());
     }
 
     @Test
     public void testGetType() {
-        Ingredient ingredient = new Ingredient(SAUCE, "Сырный", 100);
-        assertEquals(IngredientType.SAUCE, ingredient.getType());
+        Ingredient ingredient = new Ingredient(type, name, price);
+        assertEquals(type, ingredient.getType());
     }
 }
